requirements.txt
# Python Dependencies for Snowflake AI & Cortex Demo
# Install with: pip install -r requirements.txt

# Snowflake connector and pandas integration
snowflake-connector-python==3.12.0
snowflake-snowpark-python==1.21.0

# Machine Learning (for Snowpark ML pipeline)
scikit-learn==1.4.0
xgboost==2.0.3

# Data manipulation
pandas==2.1.4
numpy==1.26.3

# HTTP requests for API calls
requests==2.31.0

# HTML parsing for SEC EDGAR
beautifulsoup4==4.12.3
lxml==5.1.0

# Optional: For faster data loading
pyarrow==14.0.2

# Optional: Snowflake ML (model registry, feature store)
# Uncomment if using advanced Snowpark ML features
# snowflake-ml-python==1.2.0
